+++ container-engine-lib/lib/backend_impls/kubernetes/object_attributes_provider/label_value_consts/label_value_consts.go
@@ -20,7 +20,7 @@ const (
 	userServiceKurtosisResourceTypeLabelValueStr       			= "user-service"
 	networkingSidecarKurtosisResourceTypeLabelValueStr 			= "networking-sidecar"
 	moduleKurtosisResourceTypeLabelValueStr            			= "module"
-	filesArtifactExpanderKurtosisResourceLabelValueStr 			= "files-artifact-expander"
+	filesArtifactExpansionKurtosisResourceLabelValueStr 			= "files-artifact-expander"
 
 	enclaveDataVolumeTypeLabelValueStr = "enclave-data"
 	filesArtifactExpansionVolumeTypeLabelValueStr = "files-artifact-expansion"
@@ -43,7 +43,7 @@ var ModuleKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.Must
 var EnclaveKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(enclaveKurtosisResourceTypeLabelValueStr)
 var APIContainerKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(apiContainerKurtosisResourceTypeLabelValueStr)
 var UserServiceKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(userServiceKurtosisResourceTypeLabelValueStr)
-var FilesArtifactExpanderKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(filesArtifactExpanderKurtosisResourceLabelValueStr)
+var FilesArtifactExpansionKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(filesArtifactExpansionKurtosisResourceLabelValueStr)
 var NetworkingSidecarKurtosisResourceTypeKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(networkingSidecarKurtosisResourceTypeLabelValueStr)
 var NetworkPartitioningEnabledKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(trueValueStr)
 var NetworkPartitioningDisabledKubernetesLabelValue = kubernetes_label_value.MustCreateNewKubernetesLabelValue(falseValueStr)